+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 ### Added
+### Fixed
+
+## [7.7.1] - 2026-05-19
+### Added
 - API now provides up to date keystone timers live when computing keystone result, with hardcoded timers as a fallback.
 - Add more description to the video encoder field.
 - Add a warning when using H265/AV1 without hardware rendering.
 - Enable the "followRedirects" flag on the guild websocket.
 
-### Fixed
 
 ## [7.7.0] - 2026-05-12
 ### Added